fix feature login and logout with cookie

+++ src/modules/v1/products/product_handler.go
@@ -43,6 +43,11 @@ func (h *handler) CreateProduct(c echo.Context) error {
 		return err
 	}
 
+	role := c.Response().Header().Get("user_role")
+	if role != "admin" {
+		return c.JSON(http.StatusUnauthorized, "You are not authorized to create a product")
+	}
+
 	data, err := h.service.Insert(&product)
 	if err != nil {
 		return err
@@ -59,6 +64,11 @@ func (h *handler) UpdateProduct(c echo.Context) error {
 		return err
 	}
 
+	role := c.Response().Header().Get("user_role")
+	if role != "admin" {
+		return c.JSON(http.StatusUnauthorized, "You are not authorized to update a product")
+	}
+
 	res, err := h.service.Update(id, &product)
 	if err != nil {
 		return c.JSON(http.StatusNotAcceptable, err.Error())
@@ -70,6 +80,12 @@ func (h *handler) UpdateProduct(c echo.Context) error {
 
 func (h *handler) DeletProduct(c echo.Context) error {
 	id := c.Param("id")
+
+	role := c.Response().Header().Get("user_role")
+	if role != "admin" {
+		return c.JSON(http.StatusUnauthorized, "You are not authorized to delete a product")
+	}
+
 	res, err := h.service.Delete(id)
 	if err != nil {
 		return c.JSON(http.StatusNotAcceptable, err.Error())
